Completed basic implementation of core Opacity protocol, minus the host-side caching of accepted card IDs.

+++ code/applets/src/uk/ac/cam/bo271/applets/challengeresponse/ChallengeResponse.java
@@ -56,7 +56,7 @@ public class ChallengeResponse extends Applet {
 	private void setArray(APDU apdu) {
 		byte[] buffer = apdu.getBuffer();
 		byte Lc = buffer[ISO7816.OFFSET_LC];
-		updateBuffer((byte)Lc);
+		updateBuffer(Lc);
 		Util.arrayCopy(buffer, ISO7816.OFFSET_CDATA, byteArray, (short)0, Lc);
 	}
